refactor(models): refactor imports TimestampMixin, was circular impost error

+++ app/models/profile.py
@@ -3,7 +3,8 @@ from typing import TYPE_CHECKING
 from sqlalchemy import String, Integer, ForeignKey, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from app.models import Base, TimestampMixin
+from app.models import Base
+from app.models.mixin import TimestampMixin
 
 
 if TYPE_CHECKING:
@@ -37,7 +38,7 @@ class Profile(Base, TimestampMixin):
     )
 
     age: Mapped[int] = mapped_column(
-        Integer(120),
+        Integer,
         nullable=True,
     )
